Move ConfigurationDataLoaderExtensions to own file
